test: Move files to domain folder
